Create README.md
Tidied up from the usual marker pack in circulation, which is just a data dump of everyone's individual user pins. The aim of this pack is to consolidate and tidy markers in to relevantly named and logical files, remove double-ups and correct mistakes or omissions. In addition most of the marker graphics have been remade to be clearer and neater.